@@ -1,13 +1,63 @@
 {
   "name": "simplengrams",
-  "version": "2.0.2",
+  "version": "3.0.0",
   "description": "The easiest way to get n-gram chunks from strings or token arrays!",
-  "main": "index.js",
-  "files": [
-    "index.js",
-    "index.d.ts"
-  ],
-  "types": "index.d.ts",
+  "main": "dist/cjs/index.min.js",
+  "module": "dist/esm/index.min.js",
+  "browser": "dist/iife/index.min.js",
+  "types": "dist/cjs/index.d.ts",
+  "type": "module",
+  "directories": {
+    "dist": "dist",
+    "doc": "docs",
+    "test": "test"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.12.10",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-typescript": "^7.12.7",
+    "@rollup/plugin-babel": "^5.2.2",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.1",
+    "@types/chai": "^4.2.14",
+    "@types/node": "^14.14.14",
+    "@typescript-eslint/eslint-plugin": "^4.10.0",
+    "@typescript-eslint/parser": "^4.10.0",
+    "chai": "^4.2.0",
+    "eslint": "^7.16.0",
+    "esm": "^3.2.25",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.35.1",
+    "rollup-plugin-dts": "^2.0.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.29.0",
+    "ts-node": "^9.1.1",
+    "tslib": "^2.0.3",
+    "typedoc": "^0.19.2",
+    "typescript": "^4.1.3"
+  },
+  "scripts": {
+    "prebuild": "npm run clean:pre && npm run lint:fix",
+    "build": "npm run build:types && rollup -c rollup.config.js",
+    "postbuild": "npm run clean:post && npm run check:es5",
+    "build:docs": "typedoc --out docs",
+    "build:types": "tsc --emitDeclarationOnly --declaration --declarationDir types",
+    "clean": "npm run clean:pre && npm run clean:post && npm run clean:docs",
+    "clean:docs": "rimraf docs",
+    "clean:pre": "rimraf build && rimraf dist && rimraf types",
+    "clean:post": "rimraf types",
+    "lint": "eslint 'src/**/*.ts'",
+    "lint:fix": "eslint --fix 'src/**/*.ts'",
+    "check:types": "tsc --noEmit",
+    "check:es5": "npx es-check es5 ./dist/iife/*.js --verbose",
+    "check:es3": "npx es-check es3 ./dist/iife/*.js --verbose"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/phugh/simplengrams"
+  },
   "keywords": [
     "n-grams",
     "ngrams",
@@ -23,21 +73,12 @@
     "email": "peter@phugh.es",
     "url": "https://www.phugh.es"
   },
-  "license": "CC-BY-NC-SA-3.0",
-  "engines": {
-    "node": ">=8.10.0"
-  },
-  "scripts": {
-    "test": "node ./test/test.js"
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/phugh/simplengrams/issues"
   },
   "homepage": "https://github.com/phugh/simplengrams",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/phugh/simplengrams"
-  },
-  "devDependencies": {
-    "eslint": "^7.0.0",
-    "eslint-config-google": "^0.14.0"
-  },
-  "dependencies": {}
+  "dependencies": {
+    "core-js": "^3.8.1"
+  }
 }
